pyproject.toml
[tool.poetry]
name = "wagtail-bikeshed"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
wagtail = "^2.9.1"
psycopg2 = "^2.8"
consoler = "^0.1.2"
django-assets = "^2.0"
libsass = "^0.20.0"
django-autocomplete-light = "^3.5.1"
wagtail-factories = "^2.0.0"
black = "^19.10b0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
factory-boy = "^2.12.0"
pytest-cov = "^2.10.0"
pytest-django = "^3.9.0"
pytest-factoryboy = "^2.0.3"
pytest-pythonpath = "^0.7.3"
pytest-sugar = "^0.9.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
